requirements.txt
asgiref
astroid
certifi==2023.5.7
charset-normalizer==2.1.1
django==4.2
django-allauth
django-crispy-forms==1.14.0
django-heroku==0.3.1
gunicorn==20.1.0
idna==3.4
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.6.1
oauthlib==3.2.2
pylint==2.14.0
Pillow==9.3.0
PySocks==1.7.1
python3-openid==3.2.0
pytz==2023.2
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
stripe==5.1.0
python-dotenv
